power: Add slow charging functionality in Adaptive Charging
This adds a slow charging feature to Adaptive Charging that limits the charge current used to charge a battery from the set battery sustain percentage to 100%. Currently, Adaptive Charging charges the battery as quickly as possible to 100% just before the predicted time of when a user is likely to unplug their device. Quick-charging of the battery is not ideal as it promotes battery degradation. By starting the charge earlier and charging over a longer duration, the charge current to the battery can be decreased and this can improve long-term battery health.

BUG=b:260359049
TEST=cros_workon_make --board=brya power_manager --test

Change-Id: If330dc446ab8fec0856eab4374c2edd8e797539f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4128281
Tested-by: Jolene Loo <joleneloo@google.com>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Jolene Loo <joleneloo@google.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>