camera: disable effects based on config booleans
The effects pipeline is still checking the effects enumeration to
determine whether there are any effects set, but this is deprecated.

This CL changes the test to check the three *_enabled booleans to
determine if an effect is set.

This CL also removes an unused function to reduce the references to the
deprecated enum.

BUG=b:264204503,b:263910410
TEST=deploy to brya, change flags and check with camera app

Change-Id: I141bc0790b4bfaa875acf2191450359652da358b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4128283
Tested-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Jim Pollock <jmpollock@chromium.org>
Reviewed-by: Tom Shafron <shafron@google.com>
Commit-Queue: Jim Pollock <jmpollock@chromium.org>