vm_tools: Add {Disk,Swap}{Reads,Writes} metric handling for termina.
In much the same way as Borealis reporting KB of writes and reads to
disk and swap, we would like to be able to report the same metrics for
crostini/termina. Since the heavy lifting of information
'transportation' has already been implemented, this is a relatively
small change.

BUG=chromium:1193485
TEST=FEATURES=test emerge-volteer vm_host_tools

Change-Id: I65568c166d4b42fdfbe7aed8f55785010055da94
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4137190
Reviewed-by: Phillip Pearson <philpearson@google.com>
Reviewed-by: Nicholas Verne <nverne@chromium.org>
Commit-Queue: Maciek Swiech <drmasquatch@google.com>
Tested-by: Maciek Swiech <drmasquatch@google.com>