cryptohome: Don't do b/261188092
Hot take CL: the answer to these TODOs is to do nothing.

In one case there's nothing really necessary; if we get a failure at
that point then nothing has been written out yet, so there's no recovery
or rollback necessary.

In the other case, it's more interesting. We could try to revert any
migration bits and migrated factors if we discover that the USS file is
missing. However, I would say that this is the wrong place to handle
this; if we want to do rollback when that file is removed, we should do
so in the code that's removing it (which today does not exist). The only
way this currently should happen right now is if something external has
gone and removed the file, which I don't believe is a problem we can
really reasonably solve.

BUG=b:261188092
TEST=built code

Change-Id: Ib4785550948b4cece605e57eed435c5d7d63831a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4137209
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>