system_api: Remove key_label_data field
This removes the deprecated key_label_data field from the StartAuthSession
reply message. The cryptohome code that actually populates the field has
already been removed and so this is just removing the definition.

The client-side code which uses this is already disabled and is in the
process of being removed, although this CL will wait on that being
committed before it goes in itself.

BUG=b:261030151
TEST=ran all cryptohome unit tests with the client libs rebuilt

Change-Id: I9c54bd3a7fe5235883099654b716942580f7786f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4140423
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>