missive: Require generation id in file path.
Delete any files in StorageQueue that don't have a generation id in their path/name.
We allowed files to not have generation id in their path for legacy
reasons, but we are far enough along in versions that most systems
should have a version of Missive that stores generation ids as part of
the file path.

BUG=b:264599257
TEST=FEATURES="test" emerge-${BOARD} missive

Change-Id: Ib226cef2a233b6f440eefb32d21a1cea3cf37afa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4140428
Tested-by: Josh Hilke <jrhilke@google.com>
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Hong Xu <xuhong@google.com>