cryptohome: Hoist Initialize into the factory
This change pulls all the work of AuthSession::Initialize up into the
AuthSession::Create factory. From the point of view of callers of Create
(i.e. all non-test stuff) this doesn't really change anything, other
than reordering a few things a little bit.

For tests this is intended to make it a little easier to set up "real"
AuthSession objects without necessarily having to mock as much out. In
particular, instead of having to mock out user existence and the auth
factor loading (or manually force-injecting auth factors into the
session after the fact) you can just skip the factory and pass all those
values in directly.

One added benefit of this is that even tests which are constructing
AuthSession directly will still be using a properly initialized session,
unlike today where the tests that skip Initialize are basically using
the object in a "non standard" state.

Note that this change only applies minimal changes to the unit tests
themselves, and so it doesn't actually take advantage of this all that
much. This is because I wanted to avoid doing major changes to tests at
the same time I was rearranging the construction, to minimize the risk
of the refactor introducing a regression which was masked by the
modifications to all the tests. Cleanups of the unit tests will happen
in subsequent CLs.

BUG=b:263259501
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I140b5d259c8624cb4cf925a6faa97826fc21695e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4144444
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>