secanomalyd: decreases memfd baseline cap to one
The UMA metric for the count of successful memfd_create system calls is
only a baseline for the memfd execution detection, as such it is
sufficient to upload only one metric when this event occurs. Based on
the current limit of ten, we observe that there are more than enough
occurrences in the wild for this metric to still serve its purpose even
with the limit of one emission per execution of secanomalyd.
To save performance, the log scanning logic is modified such that once a
baseline metric has been emitted, new log lines are not parsed for this
type of event (but are still parsed for the target anomaly).

BUG=b:264702502
TEST=A single baseline metric is uploaded when a successful memfd_create
syscall is detected in the logs. Further events do not result in
additional emissions.

Change-Id: Ie46b2426e9a30706fb352510aa0f0e27775eae87
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4144450
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>