common-mk: Fix check-mojom when rename files
When we rename multiple mojom files, the old mojom could import other
old mojom. In this case, we need to also read and pass the content as
the delta info to the check script. Otherwise, it could try to read the
old file and failed because the file doesn't exist.

BUG=None
TEST=Test with CL:4145180

Change-Id: I86fe01d19be67a6bf8a7d78365cb0da0b1de9ba1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4145179
Tested-by: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>