vm_tools: cros_im: Fix typing in VS Code terminal
Currently keypresses in VS Code's terminal will usually get duplicated.
This appears to be as our implementation directly commits characters
without triggering keypress events, so this CL changes our behaviour of
handling keysym events to always send a fake key event then commit when
we receive the FilterKeypress call.

This change means that Chromium now sends a keydown event prior to the
keypress event, which matches the behaviour if cros_im is not used. It
is likely that some JS code in VS code doesn't expect the keypress
without a corresponding keydown. (As per b/264949241, there are still
other differences in events dispatched by Chromium.)

This CL also makes the fake key events generated contain the correct
modifiers bit-mask as Chromium/Electron apps seem to sometimes mangle
the fake key events and change the key code otherwise. For example when
setting a Danish keyboard layout, alt+s is 'ß', but without the modifier
bitmask set, the event's key code gets mangled to be a 's'.

The GtkPopoverWindowTest.KeySym test would be flaky with this change,
and as I have not been able to work out how to fix it, I have removed it
for now. Currently this test would have the keysym handler immediately
trigger a commit events, but with this change they now dispatch key
events, which potentially races with some other activation of the text
field or popover widget.

BUG=b:255273154
TEST=type in VS code terminal, type with modifier keys in various apps,
../test/run_tests.py

Change-Id: Ie582a89311a32efd93fd9be8906efb841be6848d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4145195
Tested-by: Timothy Loh <timloh@chromium.org>
Commit-Queue: Timothy Loh <timloh@chromium.org>
Reviewed-by: Sophia Lin <sophialin@google.com>