thinpool_migrator: Add stateful metadata generator
The stateful metadata generator is responsible for
generating:

1) Volume group configuration files: These are used
to restore the logical volume configuration during
migration.

2) Thinpool Logical Volume mapping: This represents
what extents belong to a given logical volume.

The format of the this metadata is ChromeOS specific:
it is assumed that StatefulMetadata will be used on
the block device associated with a stateful partition.

Given a device and its size, StatefulMetadata will
generate the metadata for migrating this device to use
a thinpool with all extents allocated to the unencrypted
logical volume.

BUG=b:232006113
TEST=unittests

Change-Id: I86231bebf97ed431038d1a0061bd175415bfbf85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4307564
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@google.com>