biod: Don't crash biod on SessionManager communication errors
In many cases assumption that SessionManager will be available is not
satisfied and generates lots of biod crash reports which are not biod
issues really. Biod crashes could easily exceeded the limit and leave
it stopped. Instead of crashing biod, which results in an attempt to
retrieve primary session once again, we should wait for
SessionStateChanged signal from SessionManager.

Crashing biod when there was a problem with retrieving primary user
informed Chromium that it should attempt to get fingerprint records.
Now that mechanism will be implemented using 'StatusChanged' signal
introduced in previous commits.

BUG=b:264848412 b:240872239
TEST=FEATURES=test emerge-hatch biod

Change-Id: If4d93c3a12e45de1a3ed769588c6ed614cacfacf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4150495
Commit-Queue: Patryk Duda <patrykd@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Patryk Duda <patrykd@google.com>
Reviewed-by: Andrea Grandi <agrandi@google.com>