secanomalyd: adds proc count UMA for memfd exec
Adds a new metric category for the number of anomalous processes on the
system. Adds a subcategory for the memfd execution attempt anomaly
to the new metric category. This subcategory shows the number of
processes on the system that attempted to execute a memfd. The
detection mechanism remains the same.
The new metric is uploaded at set intervals, dictated by
|kUmaReportInterval| and the set that holds the list of executable paths
is cleared upon successful emission of the metric.

BUG=b:265053007
TEST=The new metric is emitted when a memfd execution is attempted.
TEST=Repeated attempts launched through the same command are counted as
a single occurrence.
TEST=Attempts occurred through processes with varying launch commands
are counted as new unique attempts.
TEST=When the launch command is unknown, repeated attempts are counted
as unique attempts.
TEST=The executable file path is extracted correctly whether the launch
command has arguments or not.

Change-Id: I620fc76da92f6e8eeda23edadabc9ecb22ce5cc3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4153037
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>