shill: Remove RenewDHCPLease D-Bus method from Device
This method was only used by autotests and since all DHCP tests have
been migrated to tast tests, we should be able to remove it.

Note that the name of this method was never exposed in dbus-constants.h
so it should not be used by Chrome.

BUG=b:172227023
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: Id43cd0a20f61d8bdb4de42cc83429cc69e1d4d3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4154171
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>