diagnostics: Split delegate from executor directory
Move delegate to another directory under cros_healthd. This is because
delegate is a standalone bindary and we don't want it to link the code
from cros_healthd statically.
This change move delegate and related code to another directory, and
add `assert_no_deps` to make sure it doesn't depends on those we
don't expected.
To share code between healthd and delegate, those must be compiled as
shared library.

BUG=b:264005067
TEST=FEATURES=test emerge-brya diagnostics
TEST=(ON DUT)
  cros-health-tool diag --action=run_routine \
    --routine=led_lit_up --led_name=battery --led_color=blue

Change-Id: I639429be4cfa5a0d164749cd3cfb879afb1c2847
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4145180
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Tested-by: Chung-sheng Wu <chungsheng@google.com>