diagnostics: Move file_utils to libcros_healthd_base
Add a new directory //diagnostics/base for the basic utilities. The
target will be a shared library which can be linked to multiple
executables.
The reason not to put this in to //diagnostics/common is because that is
shared with wilco directory and contains some adaptor codes which is
hard to splitted.

BUG=b:264005067
TEST=FEATURES=test emerge-brya diagnostics
TEST=(ON DUT)
  cros-health-tool diag --action=run_routine \
    --routine=led_lit_up --led_name=battery --led_color=blue

Change-Id: I4f41138097896f96757740e5f7f9c6b6910de74c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4145338
Tested-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>