shill: vpn: Allow NotifyFailure() when stopped
VPNConnection::NotifyFailure() is supposed to be called on a connect
failure. It is possible that a connect failure happens after the object
is already in disconnecting or stopped state. Instead of checking that
at the call sites, this patch change the NotifyFailure() function
itself to allow such calls.

BUG=b:265104884
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=tast run $DUT network.VPNConnect.l2tp* network.VPNConnect.ikev2*

Change-Id: I4f3a7e2d14dcc6067cf0a4f09196816b5182fdb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4154396
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>