login: Replace cached_policy with a bool to track changes
As investigated in https://docs.google.com/document/d/1exVlnf4UzrFwhGhQzmHhLW6WB4LDuE-awS3kmaFCEzk/edit
we would like to avoid parsing again the policy data when there's no
change. And when a new policy data is set, we are sure it is different
from the previous version because the timestamp will differ.

The general idea is that we want to change the data on disk only if the new policy was set through ::Set() method. We shouldn't do any changes on disk in other cases. Otherwise we risk situations where policy is cleared because load failed, and we store the empty policy.

BUG=b:267603286
TEST=Unit tests pass

Change-Id: I126533f0ec5288c6670dc50396d9e732552d993c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4161654
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>