shill: cellular: own link RTNL event listener
Instead of relying on link events reported via DeviceInfo, create a
custom RTNL link event listener that will work as long as there is a
data connection up.

This is a prerequisite to having the multiplex support integrated in
the Cellular device, because it should monitor link events on
interfaces with a different interface index.

When relying on the DeviceInfo notification, the down event was
associated to handling the case of the device going away (as link down
is reported before link deleted), and shill would right away destroy
all services. This logic relied on the fact that no other process
should bring the network interface down, it only happened on device
removal. When using DeviceInfo there was no explicit notification for
the link deletion event, it was assumed that the full device
destruction involves destroying the Device reference stored within
DeviceInfo itself.

In this new logic with the custom listener, the interface down event
is treated differently, shill will only disconnect the device, as the
interface supposedly still exists. Only on a link deletion event all
services are destroyed.

BUG=b:269347783
TEST=CQ passes

Change-Id: I2925ee00c4724dafc0eb5a356a618809cc360804
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4215434
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>