biod: return error for enroll and auth session failures
When the StartEnrollSession or StartAuthSession returns with a session
containing an error, a brillo:Error with a message is generated in the
BiometricsManagerWrapper. This error message has the same error string
as in the returned enroll/auth sessions. The brillo:Error message can be
handled by the ChromeOS side to report failures to the UI.

BUG=b:230478501
TEST=cros_workon_make --board=hatch --test biod

Change-Id: I0e1083c13bf79c69f45c6f4e9960fb6b6632cd4a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4159938
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Istvan Nagy <iscsi@google.com>
Tested-by: Firas Sammoura <fsammoura@google.com>