cryptohome: Add a new NoDefault
This adds an alternative implementation of NoDefault (ExplicitInit) that
is aiming for the same basic concept but which provides an interface
that looks like a pointer to T instead of trying to look like a
reference to T or a T value.

The reference approach is clever, and is nice when it works, but it also
breaks down because C++ really just doesn't have a way to let you
transparently proxy. With classes it uses subclassing which means that
it just can never be used with final types, and it also then requires
template specialization and more cleverness for scalar types. This also
breaks down in a few ugly ways, e.g. I filed b:265164258 to try and get
it to play more nicely with pointers.

The more standard practice for this kind of "wrapper" type is to instead
make it look more like a pointer, for example like std::optional. Or for
a more similar example, the NoDestructor template in libchrome.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I7d0f8804c1a01b2ac4ec4c2376c8bfd24b4a23a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4159941
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>