clang-tidy: Add misc-misleading-identifier and misc-homoglyph
This CL adds misc-misleading-identifier and misc-homoglyph to the clang
tidy profile in order to enable checks for potential trojan source
attacks.

An investigation of these new lints found no existing occurences of
these lints in chromium browser code as of 2022-01-11, so noise
should be low from this, though if there were to be any findings
from this we would presumably like to know about it.

This is the ChromeOS version of this change whch aims to stay in sync
with the chromium version at crrev.com/c/4160525.

A description of these lints can be found at:
https://blog.llvm.org/posts/2022-01-12-trojan-source/

BUG=b:265059471
TEST=Analyzed chromium browser code to look for any current findings

Change-Id: I86800e50374378aa5ed60f962716944851a9a973
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4159944
Commit-Queue: Ryan Beltran <ryanbeltran@chromium.org>
Reviewed-by: George Burgess <gbiv@chromium.org>
Tested-by: Ryan Beltran <ryanbeltran@chromium.org>