diagnostics: Report TPM initialization time
This CL tries to calculate the TPM initialization time and report it to
our clients.

This metric is requested by GSC firmware team and will be sent back to
Google as an UMA metrics to evaluate if there is room for further
improvement.

BUG=b:265315639
TEST=tast run $IP health.ProbeBootPerformanceInfo
TEST=FEATURES=test emerge-brya diagnostics

Cq-Depend: chromium:4161342
Change-Id: Ifd965365ea7e40abcab06bf3d2bcc956e76455fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4160670
Commit-Queue: Kerker Yang <kerker@chromium.org>
Tested-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Byron Lee <byronlee@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>