camera: fix use-after-free crash
With CL:4160660 explicitly deleting all the StreamManipulator instances,
we now hit a use-after-free bug in GcamAeStreamManipulator. We need to
stop the file watcher before writing the cached TET values in the
destructor to prevent creating a watcher callback that'll reference the
to-be-destroyed GcamAeStreamManipulator.

BUG=None
TEST=Make sure camera service does not crash when switching modes in CCA
on Redrix.

Change-Id: Iea30d4d7af4a920d52b2ef3894ba0dc494fd564b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4169487
Reviewed-by: Ryota Okuji <okuji@chromium.org>
Tested-by: Ricky Liang <jcliang@chromium.org>
Commit-Queue: Ricky Liang <jcliang@chromium.org>