crosh: Add wireguard IPv6 support (part 1)
The WireGuard service in shill now use WireGuard.IPAddress property for
the client-side addresses (which is an address array, mixed of IPv4 and
IPv6), instead of the Address field in StaticIPProperties (which is an
IPv4 address).

This patch does the following changes for the wireguard command:
- Internally, use an array to represent the local addresses;
- When loading a service from shill, read the addresses from the new
  property, and if it is empty, fallback to the old property;
- When saving a service to shill, write the addresses to both
  properties (since Chrome is still not aware of the new property now).

This patch does not change the user interface: user still only able to
add one IPv4 address to the service. It will be changed in the follow-up
patch and unit test will also be updated there.

BUG=b:259186417
TEST=cargo test
TEST=new tast test added in https://crrev.com/c/4164638

Change-Id: Ie4b42598c1a0e98feb3b33a910c9e66b60faf798
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4169642
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>