power: Add unplug metrics for slow charging in Adaptive Charging
This updates existing metrics that are being reported to UMA for Adaptive Charging to help us analyze the effectiveness of the slow charging feature. Variations are added to Power.AdaptiveChargingMinutesToFull and Power.AdaptiveChargingBatteryPercentageOnUnplug to provide stats on normal charging, slow charging and mixed charging (combination of normal and slow charging). The Power.AdaptiveChargingDelayDelta metric is updated to use 3 hours as the correct delay when the available time for charging is 3 hours or more.

BUG=b:260359049
TEST=cros_workon_make --board=brya power_manager --test

Change-Id: Ia50188432dcb70cd7d313d1191e6b235a737ca52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4169490
Tested-by: Jolene Loo <joleneloo@google.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: David Lattimore <dml@chromium.org>
Commit-Queue: Jolene Loo <joleneloo@google.com>