trunks: try reload driver when write errors occur.
The attempts of reloading TPM device driver upon write error is
motivated by the fact that we have no good answer to this long-standing
issue that write error is sticky until reboot.

For testing the following cases are verified:
1. w/o any communication error, trunks works as expected.
2. injecting an error to /run/trunks-last-write-error, trunks works as
   expected, and the correct UMA is reported (by checking
   /var/lib/metrics/uma-events).
3. injecting an error to the mentioned file and hardcode another error
   in `tpm_handle.cc`, the correct UMA is reported, and trunks restarts
   . Also, unrecoverable error UMA isn't reported w/ multiple same
   errors.
4. injecting an error to the mentioned file and hardcode the same error
   in `tpm_handle.cc`, the correct UMA is reported, and trunks doesn't
   restart.
5. trunks can start on VM; this is to verify if the upstart script
   handles the error well, w/o aborting the start of trunks.
6. injecting an error to the mentioned file, trunks can start after
   rebinding the TPM driver on a real device.

Note: This is intended not to have unittests for any change in this CL,
for all the side effects are to those w/o unittests (main.cc,
trunks_dbus_service.cc, and tpm_handle.cc), and one will need to do
on-DUT test to make sure things don't break anyway.

BUG=b:264326384
TEST=See above.

Change-Id: I2d8ecfa5834ccda69957ea7fa5b8c7c23ca882e4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4173320
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Leo Lai <cylai@google.com>
Tested-by: Leo Lai <cylai@google.com>