shill: cellular: consolidate uplink/downlink property keys in capability
The key names in the 3GPP capability are declared in the header and
defined in the source file, do the same for the new uplink/downlink
property keys.

Also, rename them to be in line with other keys in the same class,
i.e. with the "Property" suffix.

BUG=None
TEST=cros_workon_make --board=${BOARD} --test shill

Change-Id: I11fe44f77181ba4c7fceadb884071c93d5c1e625
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4173856
Tested-by: Aleksander Morgado <aleksandermj@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Aleksander Morgado <aleksandermj@google.com>