cryptohome: Add label into AddKeyset API.
After the USS migration is completed both VaultKeyset
and the USS wrapped-main-key should be encrypted with
the same derived secrets. Otherwise on the next login
VaultKeyset decryption fails. That doesn't fail user
login, but user is not allowed to add PIN when the
VaultKeyset is not decrypted.
However the key label, which  uniquely identifies the
VK-to-be-reencrypted after the migration, is empty and
this results in skipping the reencryption of the migrated
keyset. This is caused by obtaining the label from
the old KeyData.
This CL fixes the issue by adding the label into
AddKeyset API, so that correct label is always sent
from AuthSession and reencryption of the migrated VK
is not skipped.

BUG=b:265688251
TEST=tast run DUT hwsec.CrossVersionLogin.small

Change-Id: Icba73adb355767b4fd8de29d26daa316bc4e7659
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4175445
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>