cryptohome: data_migrator: Skip migrating mtime/atime on ENOSPC error
Previously, MigrationHelper makes the entire migration as a failure when
it fails to temporarily save the mtime/atime of a file in its xattr.
However, since Ext4 has a limit on the size of xattr that can be stored
in a file, saving time on xattr can theoretically cause ENOSPC error
during the migration. This CL changes MigrationHelper so that it just
proceeds with the migration without copying mtime/atime for a file when
it fails to save the time in xattr for the file. This change applies to
all migrations that use this migration tool.

BUG=b:258402655, b:251764421
TEST=Manually test that, on ARCVM /data migration, a file with nearly 4K
TEST=xattr is migrated except for its mtime/atime without failing the
TEST=entire migration.
TEST=FEATURES=test emerge-hatch cryptohome
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v1
TEST=tast run <DUT> cryptohome.VaultMigration.fscrypt_v2

Change-Id: I4edc1bcd6a90b0ddf017129455cdaeec9765139d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4190919
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Ryo Hashimoto <hashimoto@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>