cryptohome: Handle legacy kiosk migration
This fixes a gap in the USS migration for legacy kiosk data. Currently
the migration code just takes the AuthFactorMetadata from the underlying
VaultKeyset, which usually is correct.

However, this doesn't work for legacy kiosk keysets which lack a kiosk
type and so just show up as a password keyset. In those cases we rely on
Chrome telling us that the factor is a kiosk, and we rely on Chrome to
tell us that this factor is a kiosk and not a password.

The current code treats this kiosk situation as a special case: normally
if type specified by the auth request and the type of the underlying
factor do not match AuthenticateAuthFactor will report an error, but if
the request type is kiosk then we disable that check. But this doesn't
fix up the metadata to also use kiosk metadata, and so you end up with a
mismatched type+metadata in the factor. This was not an issue in
practice when we only did authentication, because the metadata was not
needed, but now that migration is rewriting the factors this is a
problem.

To fix this, the change expands the kiosk special case to also rewrite
the metadata.

BUG=b:265067905
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I8c824186de2c9549961c0f5525a68aaf7d842084
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4177894
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>