cryptohome: Constrain UDA fuzzer amount of work
Put a reasonable upper boundary on the total number of test
"commands" executed in the userdataauth_fuzzer. This is to avoid
spurious timeout reports.

We chose this way instead of limiting the size of the input,
because for this fuzzer it's the "commands" (e.g., simulating a
D-Bus call) that are expensive, rather than the raw input size
(there can be a single command that consumes, say, 100 KB to
simulate a fake file on the disk).

BUG=b:254864841
TEST=take a 1MB timeout test case from clusterfuzz, verify cryptohome_userdataauth_fuzzer finishes it under 1 second

Change-Id: Ief5cd5bce859fdbe1a436e0e6eeb2112cb79a8ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4176897
Commit-Queue: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Maksim Ivanov <emaxx@chromium.org>