fusebox: delete obsolete D-Bus XML file
In an earlier version of Fusebox' D-Bus protocol, client and server
passed a variety of D-Bus messages. There was a long standing TODO (see
bug) to document the protocol in this XML file: both in making the
method and argument names in the XML file match the names used in
practice (instead of the one-method-does-everything "FuseBoxOperation"
placeholder) and in adding commentary (not just names and types).

It was also possible to generate type-safe C++ bindings from this XML
document, but that feature wasn't used.

In today's version of the protocol, every D-Bus message just contains
one field: an array of bytes that holds an encoded protobuf.

Documenting the protocol is better done in the
../system_api/dbus/fusebox/fusebox.proto file now. There's no need for
the XML file any more.

BUG=b:255702789
TEST=n/a

Change-Id: Id3add536da1e294491a7ed8abf23ae1c4c2b2311
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4179653
Commit-Queue: Nigel Tao <nigeltao@chromium.org>
Tested-by: Nigel Tao <nigeltao@chromium.org>
Auto-Submit: Nigel Tao <nigeltao@chromium.org>
Reviewed-by: Alexander Bolodurin <alexbn@chromium.org>