shill: Apply static search domains for IPv6
Currently StaticIPConfig of a service will only be applied to IPConfig
for IPv4 by design. For the SearchDomains field in StaticIPConfig, the
Connection class will write that to resolv.conf. Since the Connection
class only takes one IPConfig, when the network is IPv6-only,
SearchDomains will not be in IPConfig for IPv6 and thus will not be
known by the Connection class, which is problematic.

This patch adds the logic to apply static search domains also in IPv6
IPConfig, to make it taking effect on an IPv6-only network.

BUG=b:265680125
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=manually tested by:
  - Connect to an IPv6-only network.
  - `gdbus call --system --dest org.chromium.flimflam --object-path
    /service/{id} --method org.chromium.flimflam.Service.SetProperty
    "StaticIPConfig" "{'SearchDomains':<['example.com']>}"`
  - Disconnect and connect to the network again, verify that it is
    reflected in /etc/resolv.conf
  - Check that v6 IPConfigs of the shill Device contains "example.com"
    in SearchDomains.

Disallow-Recycled-Builds: test-failures
Change-Id: I7801555a29217c80e5a3412736b593dd364ab3ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4178296
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>