power: suspend_stress_test: Ignore sporadic failures caused by hwrng kthread
The Hardware Random Number Generator (hwrng) core driver has an
unfreezable kthread which can try to interact with the TPM after
the TPM i2c or SPI bus has been suspended. This can result in
suspend (test) failures in the form of a non-increasing s2idle
count or the EC detecting a 'sleep hang'. Ideally this would
be fixed in the kernel, but so far no suitable fix could be
identified.

Both failures are deemed harmless. After the detection of a
'sleep hang' the system tries to suspend again, which usually
succeeds, unless there are other suspend issues. In the s2idle
count scenario the TPM retries delay the kernel long enough
during noirq suspend that the RTC timer wakes the system up
again right after suspend was reached, so that the CPUs don't
spend any time in idle mode.

Ignore these failures as long as they only occur sporadically
(< 1/1000). If the error rate is higher consider it an actual
suspend failure.

The condition can be identified through the kernel log entry
"hwrng: no data available" in combination with a non-increasing
s2idle count or a 'sleep hang'.

BUG=b:265341704, b:223442909, b:250089565
TEST=run suspend_stress_test on herobrine with
  CRNG_RESEED_INTERVAL = HZ

Change-Id: Ifa1ea2bb66c4e158f2cd7ba13e6571710b362594
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4175439
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>