cryptohome: Delete AuthenticateAuthSession
This deletes the implementation of AuthenticateAuthSession, including
the equivalent CLI command a bunch of secondary functions which were
only being used by it and which would now be dead code.

This does not delete the actual API definitions from system_api, a
change which will be done later.

BUG=b:261090370,b:204482221
TEST=FEATURE=test emerge-$BOARD cryptohome

Change-Id: I57a292903734e165c41ac67cf2eb5aafc382b9ee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4182490
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>