crash: Add field path_hash to uploads.log
`path_hash` contains the MD5sum of the meta file path.

BUG=b:264307614
TEST=FEATURES=test emerge-${BOARD} crash-reporter
TEST=Manufacture crashes and run `crash_sender --dev
--max_spread_time=0`. Manually inspect that normal functionality works
as usual. Manually inspect uploads.log and ensure path_hash field is
present and looks normal. Manually make sure that the value of
`path_hash` is the MD5 sum of the path of the meta file.

Change-Id: I49a0d6afc5a7255d2773401067b00a6eb21bb99f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4182498
Tested-by: Hong Xu <xuhong@google.com>
Commit-Queue: Hong Xu <xuhong@google.com>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>